Fix _meta.js files
