+++ website/pages/chat.tsx
@@ -1,0 +1,3 @@
+import ShoppingChatApplication from "../src/applications/ShoppingChatApplication";
+
+export default ShoppingChatApplication;
