--- website/pages/editor/_meta.js
@@ -1,12 +1,0 @@
-export default {
-  index: {
-    title: "🛠️ Nestia Editor",
-    type: "page",
-    hidden: true,
-    display: "hidden",
-    theme: {
-      layout: "raw",
-      footer: false,
-    },
-  },
-};
