+++ website/src/styles/unset.css
@@ -1,0 +1,3 @@
+.initialDiv > * {
+  all: initial;
+}
